+++ reduct/bucket.py
@@ -9,12 +9,21 @@ from typing import (
     List,
     AsyncIterator,
     Union,
+    Dict,
 )
 
 from pydantic import BaseModel
 
+from reduct.error import ReductError
 from reduct.http import HttpClient
-from reduct.record import Record, parse_batched_records, parse_record
+from reduct.record import (
+    Record,
+    parse_batched_records,
+    parse_record,
+    Batch,
+    TIME_PREFIX,
+    ERROR_PREFIX,
+)
 
 
 class QuotaType(Enum):
@@ -57,6 +66,9 @@ class BucketInfo(BaseModel):
 
     latest_record: int
     """UNIX timestamp of the latest record in microseconds"""
+
+    is_provisioned: bool = False
+    """bucket is provisioned amd you can't remove it or change its settings"""
 
 
 class EntryInfo(BaseModel):
@@ -231,6 +243,53 @@ class Bucket:
             **kwargs,
         )
 
+    async def write_batch(
+        self, entry_name: str, batch: Batch
+    ) -> Dict[int, ReductError]:
+        """
+        Write a batch of records to entries in a sole request
+
+        Args:
+            entry_name: name of entry in the bucket
+            batch: list of records
+        Returns:
+            dict of errors with timestamps as keys
+        Raises:
+            ReductError: if there is an HTTP  or communication error
+        """
+
+        record_headers = {}
+        content_length = 0
+        for time_stamp, record in batch.items():
+            content_length += record.size
+            header = f"{record.size},{record.content_type}"
+            for label, value in record.labels.items():
+                if "," in label or "=" in label:
+                    header += f',{label}="{value}"'
+                else:
+                    header += f",{label}={value}"
+
+            record_headers[f"{TIME_PREFIX}{time_stamp}"] = header
+
+        async def iter_body():
+            for _, rec in batch.items():
+                yield await rec.read_all()
+
+        _, headers = await self._http.request_all(
+            "POST",
+            f"/b/{self.name}/{entry_name}/batch",
+            data=iter_body(),
+            extra_headers=record_headers,
+            content_length=content_length,
+        )
+
+        errors = {}
+        for key, value in headers.items():
+            if key.startswith(ERROR_PREFIX):
+                errors[int(key[len(ERROR_PREFIX) :])] = ReductError.from_header(value)
+
+        return errors
+
     async def query(
         self,
         entry_name: str,
@@ -290,9 +349,8 @@ class Bucket:
         """
         Get full information about bucket (settings, statistics, entries)
         """
-        return BucketFullInfo.parse_raw(
-            await self._http.request_all("GET", f"/b/{self.name}")
-        )
+        body, _ = await self._http.request_all("GET", f"/b/{self.name}")
+        return BucketFullInfo.model_validate_json(body)
 
     async def subscribe(
         self, entry_name: str, start: Optional[int] = None, poll_interval=1.0, **kwargs
@@ -368,7 +426,7 @@ class Bucket:
             params["limit"] = kwargs["limit"]
 
         url = f"/b/{self.name}/{entry_name}"
-        data = await self._http.request_all(
+        data, _ = await self._http.request_all(
             "GET",
             f"{url}/q",
             params=params,